backend/requirements.txt
fastapi
uvicorn
openai
google-genai
pydantic
python-dotenv
httpx